Completed unification of data structures to handle cl_khr_fp16
